Add comments for the endpoint , rename folder middleware to middlewares, Add delete post
